feat: change things in id model and move id to models and blah blah
