+++ internal/models/id.go
@@ -1,4 +1,4 @@
-package id
+package models
 
 import (
 	"crypto/rand"
@@ -14,17 +14,17 @@ import (
 )
 
 type IDModel struct {
-	ID        string `gorm:"primaryKey;size:32"` // if id_len changes, update size parameter
+	ID        *ID `gorm:"primaryKey;size:32"` // if id_len changes, update size parameter
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt gorm.DeletedAt `gorm:"index"`
 }
 
 func (i *IDModel) BeforeCreate(tx *gorm.DB) (err error) {
-	if i.ID == "" {
-		i.ID = string(generate())
+	if i.ID == nil || *i.ID == "" {
+		i.ID = fromBytes(generate())
 	} else {
-		if !validate(i.ID) {
+		if !validate(i.ID.String()) {
 			return fmt.Errorf("invalid ID")
 		}
 	}
@@ -32,27 +32,22 @@ func (i *IDModel) BeforeCreate(tx *gorm.DB) (err error) {
 	return nil
 }
 
-type ID struct {
-	IDModel
+func validate(id string) bool {
+	return len(id) == id_len && idRegex.Match([]byte(id))
 }
+
+type ID string
 
 const charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
 const id_len = 32
 
 var idRegex = regexp.MustCompile(fmt.Sprintf(`[a-zA-Z0-9]{%d}`, id_len))
 
-func validate(id string) bool {
-	return len(id) == id_len && idRegex.Match([]byte(id))
-}
-
-func newID(contents string) *ID {
-	return &ID{IDModel: IDModel{ID: contents}}
-}
-
 func fromBytes(str []byte) *ID {
-	id := make([]byte, id_len)
-	copy(id, str)
-	return newID(string(id))
+	txt := make([]byte, id_len)
+	copy(txt, str)
+	id := ID(txt)
+	return &id
 }
 
 func FromString(str string) *ID {
@@ -78,14 +73,14 @@ func RandomID() *ID {
 }
 
 func (i *ID) Bytes() []byte {
-	return []byte(i.ID)
+	return ([]byte)(*i)
 }
 
 func (i *ID) String() string {
 	if i == nil {
 		return ""
 	}
-	return i.ID
+	return (string)(*i)
 }
 
 func (i ID) MarshalJSON() ([]byte, error) {
@@ -102,7 +97,7 @@ func (i *ID) UnmarshalJSON(data []byte) error {
 		return errors.New("id: cannot unmarshal into nil pointer")
 	}
 
-	i.ID = s
+	*i = ID(s)
 	return nil
 }
