+++ internal/models/room.go
@@ -1,8 +1,6 @@
 package models
 
-import "github.com/Nykenik24/enkrypted/internal/id"
-
 type Room struct {
-	id.IDModel
+	IDModel
 	Password string `json:"-"`
 }
